+++ cli/internal/config/relayer_test.go
@@ -32,7 +32,7 @@ func TestRelayerConfig(t *testing.T) {
 		assert.Equal(t, 3*time.Second, *config.Relayer.DispatchPollInterval)
 		require.Len(t, config.Relayer.ChainOverrides, 2)
 		chain := config.Relayer.ChainOverrides[0]
-		assert.Equal(t, "0xe20BccD900Fa1B48f46F5a483d9De063b07eDFCC", config.Chains[0].EVM.ICS26Router)
+		assert.Equal(t, "0x0000000000000000000000000000000000000001", config.Chains[0].EVM.ICS26Router)
 		assert.Equal(t, 2*time.Second, *chain.TxSubmissionDelay)
 		//nolint:testifylint // exact literal from the fixture; a tolerance would mask decoding drift
 		assert.Equal(t, 1.5, *chain.EVM.GasFeeCapMultiplier)
