+++ cli/internal/config/testdata/sample.yml
@@ -19,11 +19,11 @@ chains:
     evm:
       rpc: https://ethereum-rpc.example.com
       ws: wss://ethereum-rpc.example.com
-      ics26Router: "0xe20BccD900Fa1B48f46F5a483d9De063b07eDFCC"
+      ics26Router: "0x0000000000000000000000000000000000000001"
   - chainId: "8453"
     evm:
       rpc: https://base-rpc.example.com
-      ics26Router: "0xe20BccD900Fa1B48f46F5a483d9De063b07eDFCC"
+      ics26Router: "0x0000000000000000000000000000000000000001"
 relayer:
   dispatchPollInterval: 3s
   chainOverrides:
